# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Hotel Booking Management API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /guests:
    get:
      operationId: listGuests
      tags: [guests]
      summary: List all unique guests with their booking history
      parameters:
        - name: search
          in: query
          required: false
          schema:
            type: string
          description: Search by name, email, or phone
      responses:
        "200":
          description: List of guests
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Guest"
  /guests/{email}:
    get:
      operationId: getGuest
      tags: [guests]
      summary: Get guest profile by email with full booking history
      parameters:
        - name: email
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Guest profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GuestProfile"
        "404":
          description: Not found
...
  /guest/units/{unitId}:
    get:
      operationId: getGuestUnit
      summary: Get unit info for guest portal
      tags: [guest]
      parameters:
        - name: unitId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Unit info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GuestUnit"
        "404":
          description: Not found
  /guest/requests:
    get:
      operationId: listGuestRequests
      summary: List guest requests
      tags: [guest]
      parameters:
        - name: roomId
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Requests
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GuestRequest"
    post:
      operationId: createGuestRequest
      summary: Submit a guest request
      tags: [guest]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GuestRequestInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GuestRequest"
  /guest/requests/{id}:
    patch:
      operationId: updateGuestRequest
      summary: Update guest request status
      tags: [guest]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                status:
                  type: string
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GuestRequest"
  /guest/feedback:
    post:
      operationId: createGuestFeedback
      summary: Submit guest satisfaction feedback
      tags: [guest]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GuestFeedbackInput"
      responses:
        "201":
          description: Recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GuestFeedback"
...
components:
  schemas:
...
    Booking:
      type: object
      required: [id, guestName, guestEmail, roomId, checkIn, checkOut, status, totalAmount, createdAt]
      properties:
        id:
          type: integer
        guestName:
          type: string
        guestEmail:
          type: string
        guestPhone:
          type: ["string", "null"]
        roomId:
          type: integer
        roomName:
          type: ["string", "null"]
        roomType:
          type: ["string", "null"]
        checkIn:
          type: string
        checkOut:
          type: string
        status:
          type: string
          description: "confirmed, checked-in, checked-out, cancelled"
        totalAmount:
          type: number
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
...
    Guest:
      type: object
      required: [guestName, guestEmail, totalBookings, totalSpent, lastStay, firstStay]
      properties:
        guestName:
          type: string
        guestEmail:
          type: string
        guestPhone:
          type: ["string", "null"]
        totalBookings:
          type: integer
        totalSpent:
          type: number
        lastStay:
          type: ["string", "null"]
        firstStay:
          type: ["string", "null"]
        lastRoomType:
          type: ["string", "null"]
        lastStatus:
          type: ["string", "null"]
    GuestProfile:
      type: object
      required: [guestName, guestEmail, totalBookings, totalSpent, lastStay, firstStay, bookings]
      properties:
        guestName:
          type: string
        guestEmail:
          type: string
        guestPhone:
          type: ["string", "null"]
        totalBookings:
          type: integer
        totalSpent:
          type: number
        lastStay:
          type: ["string", "null"]
        firstStay:
          type: ["string", "null"]
        lastRoomType:
          type: ["string", "null"]
        lastStatus:
          type: ["string", "null"]
        bookings:
          type: array
          items:
            $ref: "#/components/schemas/Booking"
...
    GuestUnit:
      type: object
      properties:
        unit:
          type: object
          properties:
            id:
              type: integer
            name:
              type: string
            type:
              type: string
            propertyId:
              type: integer
            propertyName:
              type: string
              nullable: true
            propertyType:
              type: string
              nullable: true
        financial:
          nullable: true
          type: object
          properties:
            dueDate:
              type: string
              nullable: true
            amountDue:
              type: number
              nullable: true
            status:
              type: string
            checkIn:
              type: string
              nullable: true
            checkOut:
              type: string
              nullable: true
    GuestRequest:
      type: object
      required: [id, roomId, type, description, status, refCode, createdAt]
      properties:
        id:
          type: integer
        roomId:
          type: integer
        unitName:
          type: string
          nullable: true
        propertyName:
          type: string
          nullable: true
        type:
          type: string
        description:
          type: string
        facilityName:
          type: string
          nullable: true
        scheduledAt:
          type: string
          nullable: true
        visitorName:
          type: string
          nullable: true
        visitorPhone:
          type: string
          nullable: true
        status:
          type: string
        refCode:
          type: string
        createdAt:
          type: string
    GuestRequestInput:
      type: object
      required: [roomId, type, description]
      properties:
        roomId:
          type: integer
        type:
          type: string
        description:
          type: string
        facilityName:
          type: string
        scheduledAt:
          type: string
        visitorName:
          type: string
        visitorPhone:
          type: string
    GuestFeedback:
      type: object
      required: [id, roomId, rating, createdAt]
      properties:
        id:
          type: integer
        roomId:
          type: integer
        rating:
          type: string
        comment:
          type: string
          nullable: true
        createdAt:
          type: string
    GuestFeedbackInput:
      type: object
      required: [roomId, rating]
      properties:
        roomId:
          type: integer
        rating:
          type: string
        comment:
          type: string
...